requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
easyocr==1.6.2
fastapi==0.85.0
fonttools==4.37.3
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.22.0
kiwisolver==1.4.4
matplotlib==3.6.0
networkx==2.8.6
ninja==1.10.2.3
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
outcome==1.2.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyclipper==1.3.0.post3
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
selenium==4.4.3
Shapely==1.8.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.20.4
tifffile==2022.8.12
tk==0.1.0
torch==1.12.1
torchvision==0.13.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
wsproto==1.2.0
